+++ libraries/RunAvgWeight/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - 2024-08-12
+- Fix #6, **float getFastAverage()** => fix **clear()**
+- Fix #4, add **float getCoefficientOfVariation()**
+- add bool return value for **clear(), addValue()**
+- update readme.md
+- update keywords.txt
+
+----
 
 ## [0.1.1] - 2024-07-01
 - update documentation